+++ NEWS.md
@@ -1,9 +1,9 @@
 # bigrquery (development version)
 
+* Now supports `dbAppendTable()` (#539) and `dbCreateTable()` (#483).
+
 * `dbWriteTable()` now correct uses the `billing` value set in the 
   connection (#486).
-
-* Now supports `dbAppendTable()` (#539).
 
 * `dbReadTable()`, `dbWriteTable()`, `dbExistsTable()`, `dbRemoveTable()`,
   and `dbListFields()` now all work with `DBI::Id()` (#537).